config: [HACK] cleanup when using `Config.with`
`Config.with` clones a new config object and overwrite properties that
bypasses the property protection its `provider` key. This change will reset
the provider if `Config.with` is invoked with a provider.

Why is this a hack? Well, the optional object passed to `with` may not
imply a new provider is created... so this ALWAYS resets the provider
forcing a new provider to be created which is bad.